application core
Taken from autoplay project.

This version uses a parent program to run the bot within a separate process. This should also allow for extensions to be added such as a Flask web UI.

Compared to the original bot structure, this version should be able to support more abstractions for cleaner code overall.