new message class to replace message/action dicts
Should provide the same functionality with easier to read code.

Similar to a linked list, the ``previous`` attribute maintains the message chain, allowing the data to be accessed if needed without adding unnecessary bloat and nesting within the payload.

The ``get_first_message`` method will return the root message in the chain, this is especially useful for generating responses. Eg, the first message's source is likely the response's intended target.